Langchain_Bitcoin/requirements.txt
# =============================================================================
# 비트코인 챗봇 v2.7 - 이중 데이터베이스 RAG 시스템 의존성
# Spring Boot + LangChain + PgVector + PostgreSQL + Redis 아키텍처
# =============================================================================

# 핵심 Python 런타임
python-dotenv>=1.0.0                    # 환경변수 관리

# =============================================================================
# 데이터베이스 계층 (이중 DB 아키텍처)
# =============================================================================

# PostgreSQL 연결 (PgVector + 일반 PostgreSQL)
psycopg2-binary>=2.9.9                  # PostgreSQL 드라이버 (컴파일된 버전)

# Redis 캐싱 시스템 (TTL 30초 다층 캐싱)
redis>=5.0.0                            # Redis 클라이언트

# =============================================================================
# AI/ML 핵심 스택 (LangChain + OpenAI)
# =============================================================================

# OpenAI API 통합 (GPT-3.5-turbo + text-embedding-ada-002)
openai>=1.12.0                          # OpenAI API 클라이언트

# LangChain 핵심 프레임워크
langchain>=0.1.0                        # LangChain 핵심
langchain-openai>=0.0.6                 # OpenAI 통합
langchain-community>=0.0.20             # 커뮤니티 확장
langchain-core>=0.1.0                   # 코어 컴포넌트
langchain-experimental>=0.0.50          # 실험적 기능

# =============================================================================
# 웹 프레임워크 (FastAPI)
# =============================================================================

# FastAPI 비동기 웹 서버 (포트 8003)
fastapi>=0.109.0                        # FastAPI 프레임워크
uvicorn>=0.27.0                         # ASGI 서버
python-multipart>=0.0.9                 # 멀티파트 폼 지원

# 데이터 검증 및 직렬화
pydantic>=2.5.0                         # 데이터 검증

# =============================================================================
# 데이터 처리 및 과학 계산
# =============================================================================

# 수치 계산 및 데이터 처리
numpy>=1.26.0                           # 수치 계산
pandas>=2.0.0                           # 데이터프레임

# =============================================================================
# 웹 스크래핑 및 외부 API 연동
# =============================================================================

# HTTP 클라이언트 및 웹 스크래핑
requests>=2.31.0                        # HTTP 요청
beautifulsoup4>=4.12.0                  # HTML 파싱

# 뉴스 수집 및 RSS 처리
feedparser>=6.0.10                      # RSS 피드 파싱
lxml>=4.9.3                             # XML/HTML 고속 파싱
newspaper3k>=0.2.8                      # 뉴스 기사 추출

# =============================================================================
# 텍스트 처리 및 임베딩
# =============================================================================

# 텍스트 임베딩 (대안 모델용)
sentence-transformers>=2.2.2            # 사전훈련된 임베딩 모델

# =============================================================================
# 차트 생성 및 시각화 (Plotly)
# =============================================================================

# Plotly 차트 생성 (1400x1200 기술분석 차트)
plotly>=5.17.0                          # 인터랙티브 차트
kaleido>=0.2.1                          # 정적 이미지 변환 (PNG)

# 이미지 처리
Pillow>=10.0.0                          # 이미지 처리

# =============================================================================
# 유틸리티 및 스케줄링
# =============================================================================

# 작업 스케줄링 (백그라운드 작업용)
schedule>=1.2.0                         # 정기 작업 스케줄링

# 비동기 처리 지원
asyncio                                  # 비동기 I/O (Python 기본 모듈)

# =============================================================================
# 개발 및 테스트 도구 (선택사항)
# =============================================================================

# 개발 도구 (개발환경에서만 필요)
# pytest>=7.4.0                         # 테스트 프레임워크
# pytest-asyncio>=0.21.0                # 비동기 테스트 지원
# black>=23.0.0                         # 코드 포맷팅
# flake8>=6.0.0                         # 린팅

# =============================================================================
# 선택적 기능 
# =============================================================================

# 멀티모달 기능 (필요시 주석 해제)
# pytesseract>=0.3.10                   # OCR (이미지 텍스트 추출)
# opencv-python>=4.8.0                  # 컴퓨터 비전 (설치 문제 시 제외)
# speech-recognition>=3.10.0            # 음성 인식 (OpenAI Whisper 대체)

# 웹 UI (Streamlit은 현재 사용하지 않음)
# streamlit>=1.31.0                     # 웹 UI (사용 안함)

# =============================================================================
# 버전 호환성 참고사항
# =============================================================================

# Python 버전: 3.9+ (권장: 3.13)
# 중요: psycopg2-binary는 개발용, 프로덕션에서는 psycopg2 권장
# 중요: kaleido는 시스템별 추가 설치가 필요할 수 있음
#   - macOS: brew install cairo pango gdk-pixbuf libffi
#   - Ubuntu: apt-get install libcairo2-dev libpango1.0-dev
#   - Windows: Visual C++ Redistributable 필요

# =============================================================================
# 설치 명령어
# =============================================================================

# 기본 설치:
# pip install -r requirements.txt

# 가상환경 권장:
# python3 -m venv venv
# source venv/bin/activate  # Linux/macOS
# venv\Scripts\activate     # Windows
# pip install -r requirements.txt

# =============================================================================
# 현재 시스템 사용 현황
# =============================================================================

# ✅ 실제 사용 중인 패키지:
# - fastapi, uvicorn: main.py FastAPI 서버
# - psycopg2-binary: dual_db_service.py 이중 DB 연결
# - redis: redis_price_service.py 가격 캐싱 (TTL 30초)
# - openai, langchain-*: chatbot_agent.py Custom Agent
# - plotly, kaleido: upbit_chart_generator.py 차트 생성
# - feedparser: news_collector.py RSS 수집
# - requests: API 호출 (Upbit, NewsData.io)
# - numpy, pandas: 데이터 처리 및 기술분석

# 🎯 핵심 기능별 의존성:
# - 이중 DB RAG: psycopg2-binary, openai, langchain-*
# - Redis 캐싱: redis
# - Plotly 차트: plotly, kaleido
# - 뉴스 수집: feedparser, requests, beautifulsoup4
# - FastAPI 서버: fastapi, uvicorn, pydantic